feat(ui): Email Attachments
This is a rough walkthrough of thinking about email
attachments in a "modular" fashion.

Basic idea is that Maxwell or/and Modular will look at the
type of email attachment and select & compose the
best module to preview that given attachment.

I've created a basic Attachment Data Model where
the "type" of attachment is specified, and a example
Youtube Preview "module" (which is still just a widget
for now) to illustrate the workflow.

Furthermore, I've added a "fallback"
module that renders attachments that might have
no corresponding module with a simple icon
(much like Inbox/Gmail).

To align ourselves with the UX exploration on Surfaces,
the EmbeddedSurface class represents a module socket
that is manifested as a tightly controlled embedded child
widget.

The parent module passes "requirements" such as the
module data/payload, given capabilities (ex. network,
audio...) and size constraints to the EmbeddedSurface.

The EmbeddedSurface calls the module resolver with these
given parameters and will render the resolved module.

Change-Id: Ie0be4f8e1112afa439925a3494bf8c518ad768f9